Remove the select all function for now
